requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
playwright>=1.40.0

# AI / LLM
anthropic>=0.40.0

# Data Processing
pyyaml>=6.0.1
python-dotenv>=1.0.0

# PDF Validation (optional — falls back gracefully if not installed)
PyPDF2>=3.0.0

# Scheduler
apscheduler>=3.10.0
croniter>=2.0.0

# CLI & Display
rich>=13.7.0
click>=8.1.0
tabulate>=0.9.0
